fix!: AuthMiddleware の再実装
